Move isolated code for parsing package files to shared
