+++ src/extension/sdk/utils.ts
@@ -1,9 +1,9 @@
 import * as fs from "fs";
 import * as path from "path";
 import { commands, ExtensionContext, window } from "vscode";
-import { PackageMap } from "../../debug/package_map";
 import { analyzerSnapshotPath, dartExecutableName, dartPlatformName, dartVMPath, DART_DOWNLOAD_URL, flutterExecutableName, flutterPath, FLUTTER_CREATE_PROJECT_TRIGGER_FILE, FLUTTER_DOWNLOAD_URL, FLUTTER_STAGEHAND_PROJECT_TRIGGER_FILE, isWin, showLogAction } from "../../shared/constants";
 import { Logger } from "../../shared/interfaces";
+import { PackageMap } from "../../shared/pub/package_map";
 import { flatMap, isDartSdkFromFlutter } from "../../shared/utils";
 import { findProjectFolders, hasPubspec } from "../../shared/utils/fs";
 import { fsPath, getDartWorkspaceFolders, openInBrowser } from "../../shared/vscode/utils";
